+++ backend/src/main/java/com/gamerparadise/dao/DTO/OrderItemDAODTO.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderItemDAODTO {
-    private long orderItemId;
-    private long orderId;
-    private long productId;
+    private Long orderItemId;
+    private Long orderId;
+    private Integer productId;
     private String productName;
     private String franchiseName;
     private Integer purchasePrice;
-    private int quantity;
+    private Integer quantity;
 }
